+++ dev-client/src/types/projects.types.ts
@@ -1,3 +1,5 @@
+import {UserPreview} from './users.types';
+import {SitePreview} from './sites.types';
 export type ProjectPreview = {
   id: number;
   name: string;
@@ -18,40 +20,9 @@ export type Project = {
   users: UserPreview[];
 };
 
-export type UserPreview = {
-  name: string;
-  id: number;
-};
-
-export type Role = 'member' | 'manager';
-
 export type ProjectPrivacy = 'private' | 'public';
-
-export type UserProfile = {
-  firstName: string;
-  lastName?: string;
-  role: Role;
-  id: number;
-};
 
 export type ProjectInputs = {
   units: 'imperial' | 'metric';
   source: 'survey' | 'soilgrids';
-};
-
-export type SiteDisplay = {
-  lat: number;
-  lon: number;
-  name: string;
-  id: number;
-};
-
-export type SitePreview = {
-  id: number;
-  name: string;
-  lastModified: {
-    user: UserPreview;
-    date: string;
-  };
-  percentComplete: number;
 };
