+++ dev-client/src/types/sites.types.ts
@@ -1,0 +1,17 @@
+import {UserPreview} from './users.types';
+export type SiteDisplay = {
+  lat: number;
+  lon: number;
+  name: string;
+  id: number;
+};
+
+export type SitePreview = {
+  id: number;
+  name: string;
+  lastModified: {
+    user: UserPreview;
+    date: string;
+  };
+  percentComplete: number;
+};
