+++ dev-client/src/types/users.types.ts
@@ -1,0 +1,13 @@
+export type UserPreview = {
+  name: string;
+  id: number;
+};
+
+export type Role = 'member' | 'manager';
+
+export type UserProfile = {
+  firstName: string;
+  lastName?: string;
+  role: Role;
+  id: number;
+};
